requirements.txt
streamlit==1.31.0
pymupdf==1.23.26
python-dotenv==1.0.0
Pillow==10.2.0
requests==2.31.0